requirements.txt
# Updated PyTorch for Python 3.13 compatibility
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Video processing
moviepy==1.0.3
imageio==2.35.0
imageio-ffmpeg==0.4.8

# Computer vision and image processing
opencv-python==4.8.1.78
pillow==10.0.0

# Scientific computing
numpy==1.24.3
scipy==1.11.3
scikit-image==0.21.0
scikit-learn>=1.3.0
pandas>=1.5.0
matplotlib>=3.6.0
seaborn>=0.11.0

# Web framework
streamlit==1.28.1

# AI/ML models
transformers==4.35.0

# API integrations
python-dotenv>=0.19.0
openai>=1.0.0
requests>=2.28.0

# Report generation
reportlab>=3.6.0

# Optional video encoding enhancement
ffmpeg-python>=0.2.0


streamlit>=1.28.0
opencv-python>=4.8.0
numpy>=1.24.0
pillow>=10.0.0
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
openai>=1.3.0
requests>=2.31.0
python-dotenv>=1.0.0
plotly>=5.15.0
pandas>=2.0.0
scipy>=1.11.0
matplotlib>=3.7.0
reportlab>=4.0.0
